// models/types.ts
export enum NodeNftNames {
  Infantry = 'Infantry',
  Rocketeer = 'Rocketeer',
  HeavyGunner = 'Heavy Gunner',
  Sniper = 'Sniper',
  "K9" = 'K9',
  // "DogPack" = "Dog Pack",
  // "DockPack" = "Dock Pack"
}

export enum BoosterItemName {
  waterMedium = 'AMMO',
  // waterLarge = 'Large 1L',
  fertilizeMedium = "Medium",
  fertilizeLarge = "Large"  
}